Rename gradle steps
