+++ src/main/kotlin/com/demonwav/mcdev/platform/bungeecord/creator/BungeeCordProjectCreator.kt
@@ -18,7 +18,7 @@ import com.demonwav.mcdev.creator.buildsystem.BuildDependency
 import com.demonwav.mcdev.creator.buildsystem.BuildRepository
 import com.demonwav.mcdev.creator.buildsystem.BuildSystem
 import com.demonwav.mcdev.creator.buildsystem.gradle.BasicGradleFinalizerStep
-import com.demonwav.mcdev.creator.buildsystem.gradle.BasicGradleStep
+import com.demonwav.mcdev.creator.buildsystem.gradle.GradleSetupStep
 import com.demonwav.mcdev.creator.buildsystem.gradle.GradleBuildSystem
 import com.demonwav.mcdev.creator.buildsystem.gradle.GradleFiles
 import com.demonwav.mcdev.creator.buildsystem.gradle.GradleGitignoreStep
@@ -118,7 +118,7 @@ class BungeeCordGradleCreator(
         return listOf(
             setupDependencyStep(),
             CreateDirectoriesStep(buildSystem, rootDirectory),
-            BasicGradleStep(project, rootDirectory, buildSystem, files),
+            GradleSetupStep(project, rootDirectory, buildSystem, files),
             setupMainClassStep(),
             setupYmlStep(),
             GradleWrapperStep(project, rootDirectory, buildSystem),
@@ -134,7 +134,7 @@ class BungeeCordGradleCreator(
         return listOf(
             setupDependencyStep(),
             CreateDirectoriesStep(buildSystem, rootDirectory),
-            BasicGradleStep(project, rootDirectory, buildSystem, files),
+            GradleSetupStep(project, rootDirectory, buildSystem, files),
             setupMainClassStep(),
             setupYmlStep()
         )
