+++ src/main/kotlin/com/demonwav/mcdev/platform/sponge/creator/SpongeProjectCreator.kt
@@ -18,7 +18,7 @@ import com.demonwav.mcdev.creator.buildsystem.BuildRepository
 import com.demonwav.mcdev.creator.buildsystem.BuildSystem
 import com.demonwav.mcdev.creator.buildsystem.BuildSystemType
 import com.demonwav.mcdev.creator.buildsystem.gradle.BasicGradleFinalizerStep
-import com.demonwav.mcdev.creator.buildsystem.gradle.BasicGradleStep
+import com.demonwav.mcdev.creator.buildsystem.gradle.GradleSetupStep
 import com.demonwav.mcdev.creator.buildsystem.gradle.GradleBuildSystem
 import com.demonwav.mcdev.creator.buildsystem.gradle.GradleFiles
 import com.demonwav.mcdev.creator.buildsystem.gradle.GradleGitignoreStep
@@ -129,7 +129,7 @@ class SpongeGradleCreator(
         return listOf(
             setupDependencyStep(),
             CreateDirectoriesStep(buildSystem, rootDirectory),
-            BasicGradleStep(project, rootDirectory, buildSystem, files),
+            GradleSetupStep(project, rootDirectory, buildSystem, files),
             mainClassStep,
             modifyStep,
             GradleWrapperStep(project, rootDirectory, buildSystem),
@@ -147,7 +147,7 @@ class SpongeGradleCreator(
         return listOf(
             setupDependencyStep(),
             CreateDirectoriesStep(buildSystem, rootDirectory),
-            BasicGradleStep(project, rootDirectory, buildSystem, files),
+            GradleSetupStep(project, rootDirectory, buildSystem, files),
             mainClassStep,
             modifyStep
         )
